Load data and solve
Updates the data loading to handle the JSON format taken from the NoSQL database. Still need to update the code so it pulls straight from the database. The solver now has the two stages. Firstly ensuring that the worst person is as best off as possible. Then, without anyone being worse than that, can we make the group better off. It also has the ability to ensure that they cannot be assigned to their L3 if they don't want to. Also, makes it so that they aren't assigned to the L3 they were assigned to for the first quarter. This will need to be changed if they use this code again next year.